refactor: 수강신청 서비스와, 대기열 서비스로 분리
